requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=10.0.0
matplotlib>=3.7.0
tqdm>=4.65.0
huggingface-hub>=0.17.0
hydra-core>=1.3.0
omegaconf>=2.3.0
einops>=0.6.0
timm>=0.9.0
transformers>=4.30.0
scipy>=1.10.0
scikit-image>=0.21.0
scikit-learn>=1.3.0
pandas>=2.0.0
git+https://github.com/facebookresearch/sam2.git  # Official SAM2 repository
jupyter>=1.0.0 